Move replaceCapability to func-alexa, make Async Refactor forEach loop to accommodate  async / await
