+++ src/routes/alexa.js
@@ -25,6 +25,7 @@ const getStateLimiter = require('../loaders/limiter').getStateLimiter;
 //const saveGrant = alexaFunc.saveGrant;
 const updateUserServices = require('../services/func-services').updateUserServices;
 const queryDeviceStateAsync = require('../services/func-alexa').queryDeviceStateAsync;
+const replaceCapability = require('../services/func-alexa').replaceCapabilityAsync;
 const saveGrantAsync = require('../services/func-alexa').saveGrantAsync;
 //const sendPageView = require('../services/ganalytics').sendPageView;
 //const sendPageViewUid = require('../services/ganalytics').sendPageViewUid;
@@ -45,6 +46,7 @@ router.get('/devices',
 		var devs = [];
 		for (let device of devices) {
 			var dev = {};
+			// Stringify endpointId
 			dev.endpointId = "" + device.endpointId;
 			dev.friendlyName = device.friendlyName;
 			dev.description = device.description;
@@ -55,7 +57,10 @@ router.get('/devices',
 			// Grab device attributes for use in building discovery response
 			var devAttributes = (device.attributes || null);
 			for (let capability of device.capabilities){
-				dev.capabilities.push(replaceCapability(capability, device.reportState, devAttributes, dev.displayCategories));
+				// Get Alexa-ified capability
+				let alexaCapability = await replaceCapability(capability, device.reportState, devAttributes, dev.displayCategories);
+				// Push to device capabilities
+				dev.capabilities.push(alexaCapability);
 			}
 			// Add specific RangeController interface
 			if (device.capabilities.indexOf('RangeController') > -1){
@@ -725,63 +730,6 @@ router.post('/authorization', defaultLimiter,
 		}
 });
 
-
-// router.post('/authorization', getStateLimiter,
-// 	passport.authenticate(['bearer', 'basic'], { session: false }),
-// 	function(req,res,next){
-// 		if (req.body.directive.payload.grant.type == "OAuth2.AuthorizationCode") {
-// 			logger.log('info', "[Alexa Authorization] Request body: " + JSON.stringify(req.body));
-// 			var messageId = req.body.directive.header.messageId;
-// 			var grantcode = req.body.directive.payload.grant.code;
-// 			// Pre-build success and failure responses
-// 			var success = {
-// 					event: {
-// 					header: {
-// 						messageId: messageId,
-// 						namespace: "Alexa.Authorization",
-// 						name: "AcceptGrant.Response",
-// 						payloadVersion: "3"
-// 					},
-// 					payload: {}
-// 				}
-// 			};
-// 			var failure = {
-// 				event: {
-// 					header: {
-// 						messageId: messageId,
-// 						namespace: "Alexa.Authorization",
-// 						name: "ErrorResponse",
-// 						payloadVersion: "3"
-// 					},
-// 					payload: {
-// 						type: "ACCEPT_GRANT_FAILED",
-// 						message: "Failed to handle the AcceptGrant directive"
-// 					}
-// 				}
-// 			};
-// 			// Save GrantCode and attempt to generate AccessToken
-// 			saveGrant(req.user, grantcode, function(grant) {
-// 				if (grant != undefined) {
-// 					res.status(200).json(success);
-// 					// requestAccessToken(req.user, function(accesstoken) {
-// 					// 	if (accesstoken != undefined) {
-// 					// 		logger.log('info', "[Alexa Authorization] Success, sending: " + JSON.stringify(success));
-// 							//res.status(200).json(success);
-// 					// 	}
-// 					// 	else {
-// 					// 		logger.log('error', "[Alexa Authorization] Failure, sending: " + JSON.stringify(failure));
-// 					// 		res.status(200).json(failure);
-// 					// 	}
-// 					// });
-// 				}
-// 				else {
-// 					logger.log('error', "[Alexa Authorization] General failure, sending: " + JSON.stringify(failure));
-// 					res.status(200).json(failure);
-// 				}
-// 			});
-// 		}
-// 	}
-// );
 ///////////////////////////////////////////////////////////////////////////
 // Functions
 ///////////////////////////////////////////////////////////////////////////
@@ -793,464 +741,6 @@ function getSafe(fn) {
         return undefined;
     }
 }
-// Replace Capability function, replaces 'placeholders' stored under device.capabilities in mongoDB with Amazon JSON
-function replaceCapability(capability, reportState, attributes, type) {
-	// BrightnessController
-	if(capability == "BrightnessController")  {
-		return {
-				"type": "AlexaInterface",
-				"interface": "Alexa.BrightnessController",
-				"version": "3",
-				"properties": {
-					"supported": [{
-						"name": "brightness"
-					}],
-					"proactivelyReported": reportState,
-					"retrievable": reportState
-				}
-			};
-	}
-	// ChannelController
-	if(capability == "ChannelController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.ChannelController",
-			"version": "3",
-			};
-	}
-	// ColorController
-	if(capability == "ColorController")  {
-		return {
-				"type": "AlexaInterface",
-				"interface": "Alexa.ColorController",
-				"version": "3",
-				"properties": {
-					"supported": [{
-						"name": "color"
-					}],
-					"proactivelyReported": reportState,
-					"retrievable": reportState
-				}
-			};
-	}
-	// ContactSensor
-	if(capability == "ContactSensor")  {
-		return {
-				"type": "AlexaInterface",
-				"interface": "Alexa.ContactSensor",
-				"version": "3",
-				"properties": {
-					"supported": [{
-						"name": "detectionState"
-					  }],
-					"proactivelyReported": reportState,
-					"retrievable": reportState
-				}
-			};
-	}
-	// ColorTemperatureController
-	if(capability == "ColorTemperatureController")  {
-		return {
-				"type": "AlexaInterface",
-				"interface": "Alexa.ColorTemperatureController",
-				"version": "3",
-				"properties": {
-					"supported": [{
-						"name": "colorTemperatureInKelvin"
-					}],
-					"proactivelyReported": reportState,
-					"retrievable": reportState
-				}
-			};
-	}
-	// InputController, pre-defined 4x HDMI inputs and phono
-	if(capability == "InputController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.InputController",
-			"version": "3",
-			"inputs": [{
-				"name": "HDMI1"
-			  },
-			  {
-				"name": "HDMI2"
-			  },
-			  {
-				"name": "HDMI3"
-			  },
-			  {
-				"name": "HDMI4"
-			  },
-			  {
-				"name": "phono"
-			  },
-			  {
-				"name": "audio1"
-			  },
-			  {
-				"name": "audio2"
-			  },
-			  {
-				"name": "chromecast"
-			  }
-			]};
-	}
-	// LockController
-	if(capability == "LockController")  {
-		return {
-				"type": "AlexaInterface",
-				"interface": "Alexa.LockController",
-				"version": "3",
-				"properties": {
-					"supported": [{
-						"name": "lockState"
-					}],
-					"proactivelyReported": reportState,
-					"retrievable": reportState
-				}
-			};
-	}
-	// MotionSensor
-	if(capability == "MotionSensor")  {
-		return {
-				"type": "AlexaInterface",
-				"interface": "Alexa.MotionSensor",
-				"version": "3",
-				"properties": {
-					"supported": [{
-						"name": "detectionState"
-						}],
-					"proactivelyReported": reportState,
-					"retrievable": reportState
-				}
-			};
-	}
-	// PercentageController
-	if(capability == "PercentageController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.PercentageController",
-			"version": "3",
-			"properties": {
-				"supported": [{
-					"name": "percentage"
-				}],
-				"proactivelyReported": reportState,
-				"retrievable": reportState
-			}
-		};
-	}
-	// PlaybackController
-	if(capability == "PlaybackController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.PlaybackController",
-			"version": "3",
-			"supportedOperations" : ["Play", "Pause", "Stop", "FastForward", "StartOver", "Previous", "Rewind", "Next"]
-			};
-	}
-	// PowerController
-	if(capability == "PowerController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.PowerController",
-			"version": "3",
-			"properties": {
-				"supported": [{
-					"name": "powerState"
-				}],
-				"proactivelyReported": reportState,
-				"retrievable": reportState
-				}
-			};
-	}
-	// RangeController | Interior and Exterior Blinds
-	if(capability == "RangeController" && (type.indexOf("INTERIOR_BLIND") > -1 || type.indexOf("EXTERIOR_BLIND") > -1)) {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.RangeController",
-			"instance": "Blind.Lift",
-			"version": "3",
-			"properties": {
-				"supported": [
-					{
-						"name": "rangeValue"
-					}
-				],
-				"proactivelyReported": true,
-				"retrievable": true
-			},
-			"capabilityResources": {
-				"friendlyNames": [
-				{
-					"@type": "asset",
-					"value": {
-						"assetId": "Alexa.Setting.Opening"
-					}
-				}
-				]
-			},
-			"configuration": {
-				"supportedRange": {
-					"minimumValue": 0,
-					"maximumValue": 100,
-					"precision": 1
-				},
-				"unitOfMeasure": "Alexa.Unit.Percent"
-			},
-			"semantics": {
-				"actionMappings": [
-				{
-					"@type": "ActionsToDirective",
-					"actions": ["Alexa.Actions.Close"],
-					"directive": {
-						"name": "SetRangeValue",
-						"payload": {
-							"rangeValue": 0
-						}
-					}
-				},
-				{
-					"@type": "ActionsToDirective",
-					"actions": ["Alexa.Actions.Open"],
-					"directive": {
-						"name": "SetRangeValue",
-						"payload": {
-							"rangeValue": 100
-						}
-					}
-				},
-				{
-					"@type": "ActionsToDirective",
-					"actions": ["Alexa.Actions.Lower"],
-					"directive": {
-						"name": "AdjustRangeValue",
-						"payload": {
-							"rangeValueDelta": -10,
-							"rangeValueDeltaDefault": false
-						}
-					}
-				},
-				{
-					"@type": "ActionsToDirective",
-					"actions": ["Alexa.Actions.Raise"],
-					"directive": {
-						"name": "AdjustRangeValue",
-						"payload": {
-							"rangeValueDelta": 10,
-							"rangeValueDeltaDefault": false
-						}
-					}
-				}
-				],
-				"stateMappings": [
-				{
-					"@type": "StatesToValue",
-					"states": ["Alexa.States.Closed"],
-					"value": 0
-				},
-				{
-					"@type": "StatesToRange",
-					"states": ["Alexa.States.Open"],
-					"range": {
-						"minimumValue": 1,
-						"maximumValue": 100
-					}
-				}
-				]
-			}
-			}
-	}
-	// RangeController
-	if(capability == "RangeController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.RangeController",
-			"version": "3",
-			"instance": "NodeRed.Fan.Speed",
-			"capabilityResources": {
-			  "friendlyNames": [
-				{
-                    "@type": "text",
-                    "value": {
-                      "text": "Fan Speed",
-                      "locale": "en-US"
-                    }
-				},
-				{
-                    "@type": "text",
-                    "value": {
-                      "text": "Position",
-                      "locale": "en-US"
-                    }
-                }
-			  ]
-			},
-			"properties": {
-			  "supported": [
-				{
-				  "name": "rangeValue"
-				}
-			  ],
-			  "proactivelyReported": reportState,
-			  "retrievable": reportState
-			},
-			"configuration": {
-			  "supportedRange": {
-				"minimumValue": 1,
-				"maximumValue": 10,
-				"precision": 1
-			  },
-			  "presets": [
-				{
-					"rangeValue": 1,
-					"presetResources": {
-					  "friendlyNames": [
-						{
-						  "@type": "asset",
-						  "value": {
-							"assetId": "Alexa.Value.Low"
-						  }
-						},
-                        {
-                          "@type": "asset",
-                          "value": {
-                            "assetId": "Alexa.Value.Minimum"
-                          }
-                        }
-					  ]
-					}
-				  },
-				{
-					"rangeValue": 5,
-					"presetResources": {
-					  "friendlyNames": [
-						{
-						  "@type": "asset",
-						  "value": {
-							"assetId": "Alexa.Value.Medium"
-						  }
-						}
-					  ]
-					}
-				  },
-				  {
-					"rangeValue": 10,
-					"presetResources": {
-					  "friendlyNames": [
-						{
-						  "@type": "asset",
-						  "value": {
-							"assetId": "Alexa.Value.Maximum"
-						  }
-						},
-						{
-						  "@type": "asset",
-						  "value": {
-							"assetId": "Alexa.Value.High"
-						  }
-						}
-					  ]
-					}
-				  }
-			  ]
-			}
-		 };
-	 }
-	// Speaker
-	if(capability == "Speaker") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.Speaker",
-			"version": "3",
-			"properties":{
-				"supported":[{
-						"name":"volume"
-					},
-					{
-						"name":"muted"
-					}
-				]}
-			};
-	}
-	// SceneController
-	if(capability == "SceneController") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.SceneController",
-			"version" : "3",
-			"supportsDeactivation" : false
-			};
-	}
-	// StepSpeaker
-	if(capability == "StepSpeaker") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.StepSpeaker",
-			"version": "3",
-			"properties":{
-				"supported":[{
-					  "name":"volume"
-				   },
-				   {
-					  "name":"muted"
-				   }
-				]}
-			};
-	}
-	// TemperatureSensor
-	if(capability == "TemperatureSensor") {
-		return {
-			"type": "AlexaInterface",
-			"interface": "Alexa.TemperatureSensor",
-			"version" : "3",
-			"properties": {
-                "supported": [
-                  {
-                    "name": "temperature"
-                  }
-                ],
-                "proactivelyReported": reportState,
-                "retrievable": reportState
-              }
-			};
-	}
-	// ThermostatController - SinglePoint
-	if(capability == "ThermostatController")  {
-		var supportedModes;
-		var hasModes = getSafe(() => attributes.thermostatModes);
-		if (attributes != null && hasModes != undefined) {
-			//supportedModes = attributes.thermostatModes;
-			supportedModes = attributes.thermostatModes.filter(function(value, index, arr){
-				// Google Home filter, remove modes that are not Alexa Compliant
-				return value != "ON";
-			});
-		}
-		else {
-			supportedModes = ["HEAT","COOL","AUTO"];
-		}
-		return {
-			"type": "AlexaInterface",
-            "interface": "Alexa.ThermostatController",
-            "version": "3",
-            "properties": {
-              "supported": [{
-                  "name": "targetSetpoint"
-                },
-                {
-                  "name": "thermostatMode"
-                }
-              ],
-			  "proactivelyReported": reportState,
-			  "retrievable": reportState
-            },
-            "configuration": {
-              "supportsScheduling": false,
-              "supportedModes": supportedModes
-			}
-		};
-	}
-};
 
 module.exports = router;
